+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/domain/entity/pageresult/TestResult.java
@@ -2,7 +2,6 @@ package com.tanaguru.domain.entity.pageresult;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonIdentityReference;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.tanaguru.domain.entity.audit.Page;
 import com.tanaguru.domain.entity.audit.TanaguruTest;