+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/runner/AuditRunnerSite.java
@@ -2,13 +2,17 @@ package com.tanaguru.runner;
 
 import com.tanaguru.crawler.TanaguruCrawlerController;
 import com.tanaguru.crawler.listener.TanaguruCrawlerListener;
+import com.tanaguru.domain.constant.EAuditLogLevel;
 import com.tanaguru.domain.entity.audit.Audit;
 import com.tanaguru.domain.entity.audit.TanaguruTest;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
 
 public class AuditRunnerSite extends AbstractAuditRunner implements TanaguruCrawlerListener {
+    private static final Logger LOGGER = LoggerFactory.getLogger(AuditRunnerSite.class);
     private TanaguruCrawlerController crawlerController;
 
     public AuditRunnerSite(
@@ -36,7 +40,12 @@ public class AuditRunnerSite extends AbstractAuditRunner implements TanaguruCraw
 
     @Override
     public synchronized void onCrawlNewPage(String url) {
-        webDriverGet(url);
+        try{
+            webDriverGet(url);
+        }catch (Exception e){
+            LOGGER.error("Error happened while auditing page {} : {}", url, e.getMessage());
+            auditLog(EAuditLogLevel.ERROR, "Error happened while auditing page " + url + " : " + e.getMessage());
+        }
     }
 
     @Override