+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/runner/factory/AuditRunnerFactoryImpl.java
@@ -80,16 +80,14 @@ public class AuditRunnerFactoryImpl implements AuditRunnerFactory {
         String basicAuthLogin = parameterStringMap.get(EAuditParameter.BASICAUTH_LOGIN).getValue();
         String basicAuthPassword = parameterStringMap.get(EAuditParameter.BASICAUTH_PASSWORD).getValue();
         String webdriverBrowser = parameterStringMap.get(EAuditParameter.WEBDRIVER_BROWSER).getValue();
-        BrowserName browserName = null;
+        BrowserName browserName;
         switch(webdriverBrowser) {
         	case CHROME:
         		browserName = BrowserName.CHROME;
         		break;
         	case FIREFOX:
-        		browserName = BrowserName.FIREFOX;
-        		break;
         	default:
-        		browserName = BrowserName.FIREFOX;    
+        		browserName = BrowserName.FIREFOX;
         }
 
         boolean enableScreenShot = Boolean.parseBoolean(parameterStringMap.get(EAuditParameter.ENABLE_SCREENSHOT).getValue());